# refactor(mrml-core): simplify codebase and improve performance