--- a/packages/mrml-core/src/helper/condition.rs
+++ b/packages/mrml-core/src/helper/condition.rs
@@ -1,6 +1,0 @@
-pub const START_CONDITIONAL_TAG: &str = "<!--[if mso | IE]>";
-pub const START_MSO_CONDITIONAL_TAG: &str = "<!--[if mso]>";
-pub const END_CONDITIONAL_TAG: &str = "<![endif]-->";
-pub const START_NEGATION_CONDITIONAL_TAG: &str = "<!--[if !mso | IE]><!-->";
-pub const START_MSO_NEGATION_CONDITIONAL_TAG: &str = "<!--[if !mso]><!-->";
-pub const END_NEGATION_CONDITIONAL_TAG: &str = "<!--<![endif]-->";