--- a/packages/mrml-core/src/helper/mod.rs
+++ b/packages/mrml-core/src/helper/mod.rs
@@ -8,5 +8,3 @@ pub mod sort;
 pub mod spacing;
 #[cfg(feature = "render")]
 pub mod style;
-#[cfg(feature = "render")]
-pub mod tag;