- mrml-cli binary size: 6.3M -> 6.2M
- mrml perf improvement: 
```
     Running benches/basic.rs (/home/me/mrml/target/release/deps/basic-fad51e37a998db58)
render button           time:   [13.650 µs 13.704 µs 13.775 µs]
                        change: [-51.894% -50.927% -49.339%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

     Running benches/template.rs (/home/me/Personal/mrml/target/release/deps/template-1402eed9fef9d1c4)
amario                  time:   [478.81 µs 480.04 µs 481.41 µs]
                        change: [-51.877% -51.599% -51.291%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe
```